+++ src/types/index.ts
@@ -2,7 +2,7 @@ export * from './event';
 export * from './item';
 export * from './localization';
 export * from './object-values';
+export * from './options';
 export * from './parsed-picker-input';
-export * from './picker-options';
 export * from './time-interval';
 export * from './year-picker';