+++ src/utils/index.ts
@@ -4,3 +4,4 @@ export * from './is-instance-of';
 export * from './pad-start';
 export * from './parse-picker-input';
 export * from './range';
+export * from './test-utils';