+++ CHANGELOG.md
@@ -9,18 +9,23 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Added
 - `/jest.config.ts`
 - Add unit tests for base `Picker` class
-- Add `Picker.thenUpdateItems()` method to group state changes before updating items
+- Add `Picker.updateAfter()` method to group state changes before updating items
 - Add `findLastIndex()` utility function
 - Add `getUniqueDayId` utility function
 - Add `startWith` property to `YearPicker`
 
 ### Changed
-- Change `DatePickle.constructor` signature
 - Use `getUniqueDayId` to guarantee uniqueness for any item of any picker
   - This is used to differentiate pages
+- Event listeners now do not trigger upon registration by default, but it can be changed
+  ```ts
+  dp.onItemsChange(callback); // Waits for next state change
+  dp.onItemsChange(callback, true); // Fires immediately
+  ```
 
 ### Removed
 - `/jest.config.js`
+- Remove `DatePickle` parent class
 
 ### Fixed
 - `MonthPicker` wrong hard-coded year used for building items